@@ -6,21 +6,21 @@
         "php": "^7.4",
         "ext-json": "*",
         "ext-pdo": "*",
-        "bensampo/laravel-enum": "^2.1",
+        "bensampo/laravel-enum": "^v3.0.0",
         "beyondcode/laravel-websockets": "^1.4",
         "cebe/php-openapi": "^1.3",
         "creativeorange/gravatar": "~1.0",
-        "cviebrock/laravel-elasticsearch": "^4.2",
-        "darkaonline/l5-swagger": "^7.0",
+        "cviebrock/laravel-elasticsearch": "^8.0.0",
+        "darkaonline/l5-swagger": "^8.0.0",
         "doctrine/dbal": "^2.10",
         "fideloper/proxy": "^4.0",
         "grohiro/laravel-camelcase-json": "~1.0",
         "hisorange/browser-detect": "^4.0",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/sanctum": "^2.0",
-        "laravel/socialite": "^4.4",
+        "laravel/socialite": "^v5.0.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
+        "laravel/ui": "^3.0",
         "marvinlabs/laravel-discord-logger": "^1.1",
         "monicahq/laravel-cloudflare": "^1.6",
         "nojacko/email-validator": "^1.1",
@@ -39,10 +39,10 @@
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.6",
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.6",
         "fzaninotto/faker": "^1.4",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^4.1",
+        "nunomaduro/collision": "^5.0",
         "phpunit/phpunit": "^9.0",
         "roave/security-advisories": "dev-master"
     },
@@ -58,12 +58,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ],
         "files": [
             "app/Helpers/permissions.php",
             "app/Helpers/strings.php",
