@@ -11,7 +11,8 @@
 |
 */
 
+use App\Http\Controllers\AboutController;
 use Illuminate\Support\Facades\Route;
 
 // TODO Utilize
-Route::get('/about/server-info', 'AboutController@serverInfo');
+Route::get('/about/server-info', [AboutController::class, 'serverInfo']);
